requirements.txt
discord.py
requests
python-dotenv
imgkit
websockets
